fix(core): split write module to satisfy RULE-003 line cap (ARCH-001)
Pure mechanical file split of crates/atm-core/src/write/mod.rs (1059 prod
lines > 1000 cap) into:
- write/pipeline.rs — PreparedWrite, send_mail/write_mail entry family,
  persisted-write preparation (449 prod lines)
- write/acknowledgement.rs — sealed atomic acknowledgement construction,
  admission, reply-target resolution (457 prod lines)
- write/mod.rs — request/outcome types + facade re-exports (166 prod lines)

No signature or body changes; all crate::write::* paths unchanged via
re-exports. Also applies RBQA-AU3-F002: write module is now pub(crate) in
lib.rs (consumed only through the ack/send facades). Both architecture
tripwires re-pointed to scan the full split module surface (facade +
submodules) so prohibited constructs cannot hide in a submodule file.

Verified: RULE-003 check passes; 444+20 atm-core tests, 63/63
boundary_enforcement tests; clippy --workspace --all-targets -D warnings
clean; sc-boundary finding set identical by identity (21 pre-existing,
zero new).
